Move shouldAddFilesUnderContention
